NOJIRA: Removed the temp cleanup.
Removed the temp directory cleanup from the script. This is causing problems with continuum because it doesn't have access to delete the files in the temp directory. The image gallery itself can take care of this already. The image gallery will have to be modified so that it doesn't complain when the temp directory already exists.